# docs/api/openapi-internal.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Data Connect Hub REST API",
...
    "version": "0.1.0",
    "license": {
      "name": "Apache-2.0"
    }
  },
  "servers": [
    {
      "url": "/api/v1alpha1/data"
    }
  ],
  "security": [
    {
      "tenantHeader": [],
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/connections": {
      "get": {
        "operationId": "listConnections",
        "summary": "List connections",
        "description": "Returns all connections accessible to the tenant identified by `x-tenant-id`. The response is a `DataConnectionResourceList`.\n",
        "tags": [
          "connections"
        ],
        "responses": {
          "200": {
            "description": "List of connection resources for the tenant",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataConnectionResourceList"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      },
      "post": {
        "operationId": "createConnection",
        "summary": "Create a connection",
        "tags": [
          "connections"
        ],
        "description": "Creates a connection that binds a connection type to credentials and connector-specific properties. The referenced connection type must already exist. Use `/test/credentials` to validate credentials before creating a connection.\nReturns `400` with `credentials_check_failed` when inline credentials do not satisfy the connection type schema. For inline credentials, a missing connection type returns `404`; for a `credentials_ref`, it returns `422`.\nAccepts two payload variants (untagged union): use `credentials_ref` to reference an existing secret, or `credentials` to supply inline credentials (the server creates the secret automatically).\n",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/DataConnection"
                  },
                  {
                    "$ref": "#/components/schemas/DataConnectionWithCreds"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Full connection resource with metadata, spec, and status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataConnectionResource"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
    "/connections/{id}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/connectionId"
        }
      ],
      "get": {
        "operationId": "getConnection",
        "summary": "Get a connection",
        "description": "Returns the full connection resource, including `metadata`, `resource`, and `status`.\n",
        "tags": [
          "connections"
        ],
        "responses": {
          "200": {
            "description": "Full connection resource with metadata, spec, and status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataConnectionResource"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "patch": {
        "operationId": "patchConnection",
        "summary": "Update a connection",
        "description": "Partially updates a connection using JSON Merge Patch (RFC 7396). Supply only the fields you want to change; omitted fields are left unchanged. To change the credentials secret reference, patch `credentials_ref`. Inline `credentials` in the patch body are silently ignored.\n",
        "tags": [
          "connections"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/merge-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/DataConnectionMergePatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Full connection resource with metadata, spec, and status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataConnectionResource"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          }
        }
      },
      "delete": {
        "operationId": "deleteConnection",
        "summary": "Delete a connection",
        "description": "Permanently deletes the connection. The credentials secret referenced by the connection is not deleted.\n",
        "tags": [
          "connections"
        ],
        "responses": {
          "204": {
            "description": "Connection deleted"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/connections/{id}/exports/secrets/{secret_name}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/connectionId"
        },
        {
          "name": "secret_name",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          },
          "description": "Name of the Kubernetes secret to create or overwrite"
        }
      ],
      "put": {
        "operationId": "exportConnection",
        "summary": "Export a connection to a Kubernetes secret",
        "description": "Flattens the connection metadata, properties, and credentials into a single Kubernetes secret with dot-notation keys (e.g. `data_connection.name`, `data_connection.properties.host`). If the secret already exists it is overwritten.\n",
        "tags": [
          "connections"
        ],
        "responses": {
          "204": {
            "description": "Secret created or updated"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/connections/{id}/readiness": {
      "parameters": [
        {
          "$ref": "#/components/parameters/connectionId"
        }
      ],
      "post": {
        "operationId": "checkConnectionReadiness",
        "summary": "Check connection readiness",
        "description": "Validates that the connection's credentials secret exists, matches the connection type schema, and that the data source is reachable. Updates the connection status accordingly.\nWhen a credentials secret cannot be read, fails schema validation, or the data-source check fails, the connection `status` is updated before the error is returned. A missing connection returns `502` without a status update. Returns `400` with `invalid_secret` or `credentials_check_failed` when the secret is missing or does not match the connection type schema. Returns `502` with `connection_check_failed` when the connection is not found or the data source check fails. Returns `500` with `status_update_failed` when the service cannot persist an updated connection status.\n",
        "tags": [
          "connections"
        ],
        "responses": {
          "204": {
            "description": "Connection is ready"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          },
          "502": {
            "$ref": "#/components/responses/BadGateway"
          }
        }
      }
    },
    "/connections/{id}/binary": {
      "parameters": [
        {
          "$ref": "#/components/parameters/connectionId"
        }
      ],
      "get": {
        "operationId": "downloadBinary",
        "summary": "Download binary data from a connection",
        "description": "Streams binary data (files, objects) from a data connection. The `path` query parameter is connector-specific (for example, an object key or a relative URL path). Only connectors that support binary reads will return data; unsupported connectors respond with `501`.\n",
        "tags": [
          "connections"
        ],
        "parameters": [
          {
            "name": "path",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "File path or object key to download",
            "example": "models/model.bin"
          }
        ],
        "responses": {
          "200": {
            "description": "Binary data stream",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                },
                "example": "UEsDBBQAAAAIAmuAAY4AAAACAAADAAAAYPIakvJrKrp5sQ0AiUtzUEgU0zE10y8uzStOASUmqTcEKkxUAAAsAQAA"
              }
            },
            "headers": {
              "Content-Disposition": {
                "schema": {
                  "type": "string"
                },
                "description": "Suggests a filename for the download",
                "example": "attachment; filename=\"model.bin\""
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "501": {
            "$ref": "#/components/responses/NotImplemented"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
    "/test/credentials": {
      "post": {
        "operationId": "testCredentials",
        "summary": "Test credentials against a data source",
        "description": "Tests the provided credentials by attempting to connect to the data source without creating any resources. You can use `/test/credentials` to test the credentials first. Once the test is successful, you can use `/connections` to create a connection.\n",
        "tags": [
          "test"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TestCredentials"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Credentials are valid"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          }
        }
      }
    },
    "/connection-types": {
      "get": {
        "operationId": "listConnectionTypes",
        "summary": "List connection types",
        "description": "Returns all connection types that are accessible to the tenant. This includes global connection types created by the operator and any custom connection types created by the tenant.\n",
        "tags": [
          "connection-types"
        ],
        "responses": {
          "200": {
            "description": "Connection type list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataConnectionTypeResourceList"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      },
      "post": {
        "operationId": "createConnectionType",
        "summary": "Create a connection type",
        "description": "Default connection types are created by the Data Connect Hub operator. You can also create connection types. Connection types created by a tenant are only accessible by that tenant.\n",
        "tags": [
          "connection-types"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DataConnectionType"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Connection type created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataConnectionTypeResource"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
    "/connection-types/{id}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/resourceId"
        }
      ],
      "get": {
        "operationId": "getConnectionType",
        "summary": "Get a connection type",
        "tags": [
          "connection-types"
        ],
        "responses": {
          "200": {
            "description": "Connection type details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataConnectionTypeResource"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "patch": {
        "operationId": "patchConnectionType",
        "summary": "Update a connection type",
        "description": "Partially updates a connection type using JSON Merge Patch (RFC 7396). Supply only the fields you want to change; omitted fields are left unchanged.\n",
        "tags": [
          "connection-types"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/merge-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/DataConnectionTypeMergePatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Connection type updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataConnectionTypeResource"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "delete": {
        "operationId": "deleteConnectionType",
        "summary": "Delete a connection type",
        "tags": [
          "connection-types"
        ],
        "responses": {
          "204": {
            "description": "Connection type deleted"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/flights": {
      "x-internal": true,
      "post": {
        "operationId": "createFlightService",
        "summary": "Register a flight service",
        "description": "Registers a new Flight gRPC service. The server probes the provided endpoint to discover supported connectors and triggers a connection-type audit. Returns 503 if the flight service is unreachable.\n",
        "x-internal": true,
        "tags": [
          "flight-services"
        ],
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FlightService"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Flight service registered",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FlightServiceResource"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "tenantHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "x-tenant-id",
        "description": "Tenant identifier. Required on all endpoints except /health."
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "OpenShift or Kubernetes bearer token. See [Authentication](#section/Authentication) for how to obtain one.\n"
      }
    },
...
    "schemas": {
...
      "DataConnectionResource": {
        "type": "object",
        "description": "A connection resource wrapping metadata, spec, and live status.\n",
        "required": [
          "metadata",
          "resource",
          "status"
        ],
        "properties": {
          "metadata": {
            "$ref": "#/components/schemas/ResourceMetadata"
          },
          "resource": {
            "$ref": "#/components/schemas/DataConnection"
          },
          "status": {
            "$ref": "#/components/schemas/DataConnectionStatus"
          }
        }
      },
      "DataConnectionResourceList": {
        "type": "object",
        "required": [
          "total_count",
          "items"
        ],
        "properties": {
          "total_count": {
            "type": "integer",
            "minimum": 0,
            "description": "Total number of connections for the tenant."
          },
          "items": {
            "type": "array",
            "description": "Connection resources for the tenant.",
            "items": {
              "$ref": "#/components/schemas/DataConnectionResource"
            }
          }
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Bad request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "code": "bad_request",
              "message": "Invalid request: missing required field 'name'"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Invalid or missing authentication token",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "code": "unauthenticated",
              "message": "invalid or missing bearer token"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Insufficient permissions",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "code": "forbidden",
              "message": "permission denied"
            }
          }
        }
      },
      "ServiceUnavailable": {
        "description": "Backend service unavailable",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "code": "service_unavailable",
              "message": "Metadata store is not available"
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "code": "not_found",
              "message": "Data connection 'abc' not found for tenant 't1'"
            }
          }
        }
      },
...
    }
  }
}